Created a seprate file for the edit profile and its feature
