+++ profile.html
@@ -12,6 +12,7 @@
     <link rel="icon" type="image/png" href="assets/favicon.ico">
     <script src="js/firebaseSetup.js" defer type="module"></script>
     <script src="js/script.js" defer type="module"></script>
+    <script src="js/profileScript.js" defer type="module"></script>
     <script src="js/darkMode.js" defer></script>
 </head>
 <body>
